requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
discord.py==1.2.3
entrypoints==0.3
flake8==3.7.8
idna==2.8
mccabe==0.6.1
multidict==4.5.2
pycodestyle==2.5.0
pyflakes==2.1.1
requests==2.22.0
soupsieve==1.9.3
urllib3==1.25.3
websockets==6.0
yarl==1.3.0
